requirements.txt
# ipwhois~=1.2.0
dnspython~=2.1.0
maxminddb-geolite2~=2018.703